kubeflow_women_clothing_ecom_reviews/pipeline/components/data_validation/requirements.txt
pyyaml==5.4.1
pandas==1.1.5
pandera==0.6.2
